requirements.txt
altair
together
python-dotenv
requests
streamlit

